scripts/requirements.txt
pymongo
python-dotenv
matplotlib
pandas